java-curl-lib/src/main/java/com/theten52/tools/network/JavaCurlPlay.java
package com.theten52.tools.network;

import com.roxstudio.utils.CUrl;

import java.util.HashMap;
import java.util.Map;

public class JavaCurlPlay {
    public static CUrl customUserAgentAndHeaders() {
        String mobileUserAgent = "Mozilla/5.0 (Linux; U; Android 8.0.0; zh-cn; KNT-AL10 Build/HUAWEIKNT-AL10) "
                + "AppleWebKit/537.36 (KHTML, like Gecko) MQQBrowser/7.3 Chrome/37.0.0.0 Mobile Safari/537.36";
        Map<String, String> fakeAjaxHeaders = new HashMap<String, String>();
        fakeAjaxHeaders.put("X-Requested-With", "XMLHttpRequest");
        fakeAjaxHeaders.put("Referer", "http://somesite.com/fake_referer");
        CUrl curl = new CUrl("http://httpbin.org/get")
                .opt("-A", mobileUserAgent) // simulate a mobile browser
                .headers(fakeAjaxHeaders)   // simulate an AJAX request
                .header("X-Auth-Token: xxxxxxx"); // other custom header, this might be calculated elsewhere
        curl.exec();
        return curl;
    }

    public static CUrl exec(String url, Map<String, String> headers) {
        CUrl curl = new CUrl(url).headers(headers);
        curl.exec();
        return curl;
    }
}
